+++ doc/CHANGES.rst
@@ -14,6 +14,7 @@ CHANGELOG
 Future
 ------
 
+* [IMP] Update Doc
 * [FIX] Remove nullable column, the nullable constraint is removed not the column
 * [ADD] Formater, convert value 2 str or str 2 value, with or without mapping
 * [ADD] CSV Importer